move PortsConfig to protos module
